+++ model/artifact.go
@@ -27,6 +27,8 @@ type Artifact struct {
 	AnyOS bool `yaml:"any-os"`
 	// if not empty use this value for the actual storage file
 	OverrideStorageBase string
+	// optionally specify the repository to archive/fetch this artifact
+	RepositoryName string `yam:"repository`
 }
 
 // StorageBase returns the file name to which the artifact is stored.