+++ CHANGES.md
@@ -1,6 +1,9 @@
 # changes
 
-###  2015-11-14
+### 2016-04-02
+- artifacts (parts) can specify their repository name
+
+### 2015-11-14
 - separate repo name from plugin
 - add support for zip compression